requirements.txt
colorama==0.4.6
mysql-connector-python==8.0.33
psycopg2-binary==2.9.6
pyyaml==6.0
